redhat: configs: revert QCOM_GENI_SE & SERIAL_QCOM_GENI to modules
Upstream-Status: RHEL-only
JIRA: https://issues.redhat.com/browse/RHEL-75936

Recent Ride4 boards needs the driver to do rely on the GearVM to do the
resource management over SCMI. What that implies is that the currently
upstream drivers for sa8775p do not work with recent BSP firmware.

Until that situation is sorted upstream, revert the host-controller
driver to a module as the vendor plans to provide an out-of-tree module
to override the upstream one in the interim.

Signed-off-by: Eric Chanudet <echanude@redhat.com>