redhat/configs: build-in SCMI power-domain driver
JIRA: https://issues.redhat.com/browse/RHEL-76561
Upstream-Status: RHEL-only

CONFIG_ARM_SCMI_POWER_DOMAIN=m -> =y

Change to the default setting and make it match kernel-ark setting. This
will be required for Qualcomm platforms that rely on firmware
initialized resources over SCMI.

Remove the chance of a deferral if built as a module and included in the
initramfs. The uncompressed kernel image size is the same for both
configuration, so there is no drawback.

Signed-off-by: Eric Chanudet <echanude@redhat.com>