redhat: Delete old changelogs to trim kernel.changelog
Upstream Status: RHEL-only
JIRA: INTERNAL

This commit removes the files with old changelogs to trim the current
kernel.changelog for automotive. If we need to check old changelogs, we
can check the git history if necessary. We just need to keep the current
development stream changelog because it is updated daily.

Signed-off-by: Julio Faracco <jfaracco@redhat.com>